+++ components/shape/src/shape.hpp
@@ -8,7 +8,7 @@ class shape : public ecs::Component
     shape(); 
     shape(Json::Value);
     Json::Value save();
-    uint16_t x, y;
-    uint16_t width, height;
+    uint32_t x, y;
+    uint32_t width, height;
     uint8_t r, g, b, a;
 };
